requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
influxdb-client==1.36.1
mysql-connector-python==8.0.32
protobuf==3.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
reactivex==4.0.4
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.14
amqp==5.0.9
anyjson==0.3.3
apturl==0.5.2
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
Brlapi==0.8.3
celery==5.2.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
elastic-transport==8.4.0
elasticsearch==8.9.0
ephem==4.1.3
evdev==1.6.1
fasteners==0.14.1
filelock==3.12.0
future==0.18.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
influxdb-client==1.37.0
jeepney==0.7.1
keyring==23.5.0
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==8.0.32
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.5.1
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-memcached==1.58
python-xlib==0.33
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reactivex==4.0.4
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.23.0
wadllib==1.3.6
wcwidth==0.2.5
xdg==5
xkit==0.0.0
zipp==1.0.0
amqp==5.0.9
anyjson==0.3.3
apturl==0.5.2
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
Brlapi==0.8.3
celery==5.2.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
elastic-transport==8.4.0
elasticsearch==8.9.0
ephem==4.1.3
evdev==1.6.1
fasteners==0.14.1
filelock==3.12.0
future==0.18.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
influxdb-client==1.37.0
jeepney==0.7.1
keyring==23.5.0
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mysql-connector-python==8.0.32
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.5.1
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-memcached==1.58
python-xlib==0.33
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reactivex==4.0.4
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.23.0
wadllib==1.3.6
wcwidth==0.2.5
xdg==5
xkit==0.0.0
zipp==1.0.0
